+++ example/lib/constants.dart
@@ -1,6 +1,8 @@
 // Dart imports:
 import 'dart:math';
 
+// Project imports:
+
 /// Note that the userID needs to be globally unique,
 final String localUserID = Random().nextInt(100000).toString();
 
@@ -14,12 +16,12 @@ enum LayoutMode {
 
 extension LayoutModeExtension on LayoutMode {
   String get text {
-    var mapValues = {
-      LayoutMode.defaultLayout: "default",
-      LayoutMode.full: "full",
-      LayoutMode.hostTopCenter: "host top center",
-      LayoutMode.hostCenter: "host center",
-      LayoutMode.fourPeoples: "four peoples",
+    final mapValues = {
+      LayoutMode.defaultLayout: 'default',
+      LayoutMode.full: 'full',
+      LayoutMode.hostTopCenter: 'host top center',
+      LayoutMode.hostCenter: 'host center',
+      LayoutMode.fourPeoples: 'four peoples',
     };
 
     return mapValues[this]!;